requirement.txt
ollama==0.4.4
chromadb==0.6.3
GitPython==3.1.44
langchain==0.3.14
langchain-text-splitters==0.3.5
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1
requests==2.32.3